refactor: move the JacksonConverter to the messaging module
