+++ messaging/src/test/java/org/axonframework/serialization/json/JacksonConverterTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.axonserver.connector;
+package org.axonframework.serialization.json;
 
 import org.junit.jupiter.api.*;
